+++ packages/ripple-binary-codec/HISTORY.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+### Fixed
+* Fix serialization/deserialization issues in `Issue` serialized type for MPTIssue.
+
+## 2.5.0 (2025-07-29)
+
 ### Added
 * Support for `Single Asset Vault` (XLS-65)
 * Adds new `STNumber` serialization type.