+++ packages/xrpl/HISTORY.md
@@ -4,19 +4,38 @@ Subscribe to [the **xrpl-announce** mailing list](https://groups.google.com/g/xr
 
 ## Unreleased
 
+### Fixed
+* Fix incorrect type checking in `validateVaultCreate` that prevented vault creation with MPT as an asset.
+
+## 4.4.2 (2025-09-25)
+
+### Fixed
+* improve Batch inner transaction typing
+
+## 4.4.1 (2025-08-29)
+
+### Fixed
+* Prevent Node.js process termination when WebSocket send() errors after the connection is closed.
+
+## 4.4.0 (2025-07-29)
+
 ### Added
+* Support for `PermissionedDEX` (XLS-81d)
+* Support for `Token Escrow` (XLS-85)
 * Support for `Single Asset Vault` (XLS-65)
 * Adds `XRPLNumber` amount type used in Vault transactions. This supports integer, decimal, or scientific notation strings.
 * Adds `ClawbackAmount` amount type used in transactions related to Clawback.
 * Fixed minified `build/xrpl-latest-min.js` to have all the latest xrpl package changes.
+* Add warning messages to `MPTokenIssuanceCreate` transaction as per [XLS-89d](https://github.com/XRPLF/XRPL-Standards/pull/293).
 
 ### Fixed
 * Fix `AccountRoot` ledger object to correctly parse `FirstNFTokenSequence` field
 * Fail faster on `tem` errors with `submitAndWait`
 * Improved type-checking in models
 * Fix issue with some transactions that would crash in validation
+* Improve typing of `Batch` inner transactions
 
-## 4.3.0 (2025-6-09)
+## 4.3.0 (2025-06-09)
 
 ### Added
 * Support for `NFTokenMintOffer` (XLS-52)